requirements.txt
typing_extensions==4.11.0
langchain
# langchain-anthropic
langchain_mistralai
langgraph
langgraph-supervisor
python-docx
pypdf